% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulhydro}
\alias{bulhydro}
\title{Donnees de base de la banque Hydro pour la Loire}
\format{Data frame with 103 observations and 27 variables

\describe{


\item{Mois.}{Debit d'un mois donne, en m3/s}
\item{Qualité.}{Niveau de validation de la mesure}

}}
\source{
\url{http://banquehydro.fr}
}
\usage{
bulhydro
}
\description{
Debit du fleuve mensuel depuis >100 ans mesuré à Saumur (code L8000020)
}
\keyword{datasets}
