% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debit_mensuel.R
\name{debit_mensuel}
\alias{debit_mensuel}
\title{Extraction tableau des debits}
\usage{
debit_mensuel(x)
}
\arguments{
\item{x}{Input table of \code{debit_mensuel} class}
}
\value{
A data frame
}
\description{
Pour les donnees banque hydro d'une station, extraire données historiques de debits mensuel
}
\details{
Depuis le dernier fichier CSV pour la station, extraire les donnees de debit
}
